requirements.txt
alembic==1.5.2
aniso8601==8.1.0
appdirs==1.4.3
attrs==20.3.0
bcrypt==3.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.6.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lockfile==0.12.2
Mako==1.1.4
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.22
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==0.16.1
